[refactor]: 서브사이드바 관련 폴더 구조 정리
