+++ frontend/src/components/ui/sidebar/subSidebar/mindmap/MindmapCard.tsx
