+++ frontend/src/components/ui/sidebar/subSidebar/pomodoro/PomodoroSubSidebar.tsx
@@ -1,6 +1,6 @@
 import { LinkIcon, Unlink, Plus } from 'lucide-react';
 import { SubSidebarAccordion } from '@/components/ui/SubSidebarAccordion.tsx';
-import CommonSubSidebarWrapper from './CommonSubSidebarWrapper';
+import CommonSubSidebarWrapper from '../CommonSubSidebarWrapper';
 import { PomodoroItem } from '@/components/ui/pomodoro/PomodoroItem.tsx';
 import type { PomodoroList } from '@/types/pomodoro';
 import AddPomodoro from '@/components/ui/Modal/AddPomodoro.tsx';
